I see this in my nightmares

+++ AirEase-AMS/App/Defs/IRoute.cs
@@ -6,7 +6,7 @@ namespace AirEase_AMS.App.Defs;
 public interface IRoute
 {
     bool FlightExists(IRoute flight);
-    int GetDistance();
+    double GetDistance();
     IGraphNode Origin();
     IGraphNode Destination();
     List<Flight>? FindFlightsInRange(DateTime begin, DateTime end);
